requirements.txt
# SOCIA Project Requirements
# 
# Note: This file has been optimized to resolve dependency conflicts
# Some packages have been commented out due to incompatible version requirements
# 
# Dependency conflicts identified:
# - google-genai requires anyio>=4.8.0 but openai requires anyio<4
# - mistralai/ag2 require httpx>=0.28.1 but may conflict with other packages
# - autogen-core requires pillow>=11.0.0 which may cause compatibility issues
#
# To use the full feature set, you may need to install some packages separately
# or create separate environments for different LLM providers

# Core scientific computing packages
numpy==1.24.3
pandas==2.0.1
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.6.1
plotly==5.15.0
networkx==3.1

# LLM and AI packages
openai>=1.1.1,<2.0.0
google-generativeai==0.8.5
# google-genai==1.11.0  # Commented out due to dependency conflicts with openai
anthropic==0.49.0
# mistralai==1.7.0  # Commented out due to dependency conflicts
transformers==4.49.0
huggingface_hub==0.29.2
datasets==3.3.2
tiktoken==0.9.0

# LangChain ecosystem
langchain==0.3.12
langchain-core==0.3.25
langchain-text-splitters==0.3.3

# AutoGen multi-agent framework (commented out due to dependency conflicts)
# ag2==0.8.7
# autogen-agentchat==0.5.3
# autogen-core==0.5.3
# autogen-ext==0.5.3
# autogenstudio==0.1.5

# Web framework and API
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2
httpx>=0.25.0,<1.0.0
httpcore>=0.18.0,<1.0.0

# Data validation and parsing
pydantic==2.11.3
pydantic-core==2.33.1

# Utility packages
tqdm==4.67.1
tenacity==9.0.0
python-dotenv==1.0.0
pyyaml==6.0
dependency-injector==4.41.0
click==8.1.8
rich==14.0.0
loguru==0.7.3

# Docker and containerization
docker==7.1.0

# Geographic and spatial analysis
geopy==2.4.1
shapely==2.0.0

# Testing
pytest==7.3.1

# Database
sqlalchemy==2.0.40
sqlmodel==0.0.24

# Async and networking
anyio>=3.7.1,<4.0.0
aiofiles==24.1.0
websockets==15.0.1

# NLP and text processing
nltk==3.9.1

# Google Cloud services
google-cloud-aiplatform==1.89.0
google-cloud-storage==2.19.0
google-api-python-client==2.167.0

# Additional utilities and dependencies
packaging>=23.2,<25.0
urllib3==2.4.0
charset-normalizer==3.4.1
requests>=2.31.0
typer==0.15.2
distro==1.9.0
orjson>=3.9.0
regex>=2023.0.0
jsonschema>=4.20.0
jsonschema-specifications>=2024.10.0
referencing>=0.36.0
rpds-py>=0.24.0

# ========================================
# OPTIONAL DEPENDENCIES (install separately if needed)
# ========================================
#
# For Mistral AI support:
# pip install mistralai==1.7.0
#
# For Google GenAI support (conflicts with OpenAI):
# pip install google-genai==1.11.0
#
# For AutoGen framework (may require pillow>=11.0.0):
# pip install ag2==0.8.7 autogen-agentchat==0.5.3 autogen-core==0.5.3 autogen-ext==0.5.3
#
# Note: Installing these packages may cause dependency conflicts with the core requirements
# Consider using separate conda environments for different LLM providers 